495: Ensured elastic search body is getting json encoded correctly
