+++ src/Api/Dto/Event.php
@@ -7,8 +7,8 @@ use ApiPlatform\Metadata\ApiProperty;
 use ApiPlatform\Metadata\ApiResource;
 use ApiPlatform\Metadata\Get;
 use ApiPlatform\Metadata\GetCollection;
-use App\Api\Filter\DateFilter;
-use App\Api\Filter\EventTagFilter;
+use App\Api\Filter\ElasticSearch\DateFilter;
+use App\Api\Filter\ElasticSearch\EventTagFilter;
 use App\Api\State\EventRepresentationProvider;
 use App\Model\DateLimits;
