+++ src/Api/Filter/ElasticSearch/DateFilter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Api\Filter;
+namespace App\Api\Filter\ElasticSearch;
 
 use ApiPlatform\Elasticsearch\Filter\AbstractFilter;
 use ApiPlatform\Metadata\Operation;
@@ -59,7 +59,7 @@ final class DateFilter extends AbstractFilter
             }
         }
 
-        return isset($ranges[1]) ? ['bool' => ['must' => $ranges]] : $ranges[0];
+        return isset($ranges[1]) ? $ranges : $ranges[0];
     }
 
     public function getDescription(string $resourceClass): array
